Moved Mojang Resolution to its own class
